requirements.txt
requests
beautifulsoup4
tqdm
python-whois
pyOpenSSL
shodan
